env example included. refactor to pull values from .env

+++ build.sh
@@ -1,5 +1,15 @@
 #! /bin/bash
 
+# read in .env values
+if [[ ! -f .env ]] ; then
+    echo "You are missing the required .env file."
+    exit 1
+fi
+source .env
+
+echo "Migrating the database"
+goose mysql "$DB_USER:$DB_PASSWORD@tcp($DB_HOST:$DB_PORT)/$DB_NAME?parseTime=true" up
+
 echo "Killing any old processes"
 killall ofc-uptime > /dev/null 2>&1
